Add files via upload
The Scratch project is basically a soccer game where you need to click the ball in order to kick it. The game is based merely on luck since the ball is programmed to go to a random position each time. The score will change by 1 if the ball is not touching the edge and not touching the goalie.